+++ src/Exceptions/RemoteException.php
@@ -6,7 +6,7 @@
  * Time: 22:44
  */
 
-namespace fortrabbit\Copy\exceptions;
+namespace fortrabbit\Copy\Exceptions;
 
 use yii\base\Exception;
