+++ src/Services/Dump.php
@@ -7,7 +7,7 @@
  * @copyright Copyright (c) 2018 Oliver Stark
  */
 
-namespace fortrabbit\Copy\services;
+namespace fortrabbit\Copy\Services;
 
 use craft\base\Component;
 use craft\helpers\FileHelper;
